requirements.txt
fastapi
uvicorn[standard]
jinja2
databases
sqlalchemy
python-dotenv
python-multipart
